+++ spec/wavefront-sdk/role_spec.rb
@@ -33,6 +33,34 @@ class WavefrontRoleTest < WavefrontTestBase
     end
   end
 
+  def test_grant
+    assert_posts("/api/v2/role/grant/#{permission}", roles.to_json) do
+      wf.grant(permission, roles)
+    end
+
+    assert_raises(Wavefront::Exception::InvalidRoleId) do
+      wf.grant(permission, [invalid_id])
+    end
+
+    assert_raises(Wavefront::Exception::InvalidPermission) do
+      wf.grant('made_up_permission', roles)
+    end
+  end
+
+  def test_revoke
+    assert_posts("/api/v2/role/revoke/#{permission}", roles.to_json) do
+      wf.revoke(permission, roles)
+    end
+
+    assert_raises(Wavefront::Exception::InvalidRoleId) do
+      wf.revoke(permission, [invalid_id])
+    end
+
+    assert_raises(Wavefront::Exception::InvalidPermission) do
+      wf.revoke('made_up_permission', roles)
+    end
+  end
+
   private
 
   def api_class